@@ -1,59 +1,76 @@
-name: android
+name: simple_material_app
 description: A new Flutter project.
 
-# The following line prevents the package from being accidentally published to
-# pub.dev using `pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: 'none'
 
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.16.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  flutter_colorpicker: ^0.5.0
+  json_annotation: ^4.0.1
+  sqflite: ^2.0.0+3
+  path_provider: ^2.0.2
+  scoped_model: ^1.1.0
+  get_it: ^1.0.3
+  firebase_core: ^1.10.0
+  firebase_database: ^9.0.6
+  flutter_spinkit: ^5.1.0
+  firebase_auth:
+  after_layout: ^1.1.0
+  cached_network_image: ^3.1.0+1
+  http: ^0.13.4
+  provider: ^6.0.1
+  shared_preferences: ^2.0.8
+  smooth_page_indicator: ^1.0.0+2
+  url_launcher: ^6.0.12
+  webview_flutter: ^3.0.0
+  charts_common: ^0.12.0
+  charts_flutter: ^0.12.0
 
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+dependencies_overrides:
+  photo_view: ^0.12.0
 
 dev_dependencies:
+  build_runner: ^2.0.5
+  json_serializable: ^4.1.3
   flutter_test:
     sdk: flutter
 
+
 # For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
+# following page: https://www.dartlang.org/tools/pub/pubspec
 
 # The following section is specific to Flutter.
 flutter:
 
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
-  # the material Icons class.
+  # the Icons class.
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
   # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+  #  - images/a_dot_burr.jpeg
+  #  - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
+  # https://flutter.io/assets-and-images/.
 
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
+  # To add assets from package dependencies, first ensure the asset
+  # is in the lib/ directory of the dependency. Then,
+  # refer to the asset with a path prefixed with
+  # `packages/PACKAGE_NAME/`. The `lib/` is implied, do not
+  # include `lib/` in the asset path.
+  #
+  # Here is an example:
+  #
+  # assets:
+  #  - packages/PACKAGE_NAME/path/to/asset
 
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
@@ -71,6 +88,3 @@ flutter:
   #       - asset: fonts/TrajanPro.ttf
   #       - asset: fonts/TrajanPro_Bold.ttf
   #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
