@@ -1,6 +1,51 @@
 import 'package:flutter/material.dart';
+import 'package:simple_material_app/pages/graph-screen.dart';
+import 'package:simple_material_app/utils/datetime-chart.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:provider/provider.dart';
+import 'package:firebase_database/firebase_database.dart';
+import 'package:simple_material_app/services/shared_preference_service.dart';
 
-void main() {
+// void main() {
+
+//   runApp(MaterialApp(
+//       // Title
+//       title: "Simple Material App",
+//       // Home
+//       home: Scaffold(
+//         // Appbar
+//         appBar: AppBar(
+//           // Title
+//           title: Text("Simple Material App"),
+//         ),
+//         // Body
+//         // body: Container(
+//         body: Center(
+//         // Center is a layout widget. It takes a single child and positions it
+//         // in the middle of the parent.
+//           child: Column(
+//           // Center the content
+//           // child: Center(
+//               // Add Text
+//               // child: GraphScreen()),
+//               mainAxisAlignment: MainAxisAlignment.center,
+//               children: <Widget>[
+//                 Text(
+//                   'You have pushed the button this many times:',
+//                 ),
+//                 Text(
+//                   '$_counter',
+//                   style: Theme.of(context).textTheme.headline4,
+//                 ),
+//               ],
+//         // ),
+//       )));
+// }
+
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
+  await SharedPreferencesService().init();
   runApp(MyApp());
 }
 
@@ -32,7 +77,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
 
   // This widget is the home page of your application. It is stateful, meaning
   // that it has a State object (defined below) that contains fields that affect
@@ -72,46 +117,36 @@ class _MyHomePageState extends State<MyHomePage> {
     // fast, so that you can just rebuild anything that needs updating rather
     // than having to individually change instances of widgets.
     return Scaffold(
-      appBar: AppBar(
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
-        title: Text(widget.title),
-      ),
-      body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
-        child: Column(
-          // Column is also a layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Invoke "debug painting" (press "p" in the console, choose the
-          // "Toggle Debug Paint" action from the Flutter Inspector in Android
-          // Studio, or the "Toggle Debug Paint" command in Visual Studio Code)
-          // to see the wireframe for each widget.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              'You have pushed the button this many times:',
-            ),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headline4,
-            ),
-          ],
+        appBar: AppBar(
+          // Here we take the value from the MyHomePage object that was created by
+          // the App.build method, and use it to set our appbar title.
+          title: Text(widget.title),
         ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: Icon(Icons.add),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
-    );
+        body: Container(
+          child: DateTimeChart(),
+        )
+        //    Center(
+        //     child: Column(
+        //       mainAxisAlignment: MainAxisAlignment.center,
+        //       children: <Widget>[
+        //         Text(
+        //           'You have pushed the button this many times:',
+        //         ),
+        //         Text(
+        //           'You have pushed the button this many times123:',
+        //         ),
+        //         Text(
+        //           '$_counter',
+        //           style: Theme.of(context).textTheme.headline4,
+        //         ),
+        //       ],
+        //     ),
+        //   ),
+        //   floatingActionButton: FloatingActionButton(
+        //     onPressed: _incrementCounter,
+        //     tooltip: 'Increment',
+        //     child: Icon(Icons.add),
+        //   ), // This trailing comma makes auto-formatting nicer for build methods.
+        );
   }
 }
